Add from_relative to Body class
Add the from_relative method of creating a Body to allow easy creation
of bodies for exoplanets and other bodies that are often described with
mass and radius relative to Sun, Jupiter or Earth.  The from_relative
method will take any object as a reference as long as it implements the
attributes (k and R) and does not require the reference object to be a
poliastro Body object.

Updated the docstring on Body to remove reference to the Solar System
as bodies may be exoplanets, stars or other extrasolar objects.

Allowed for pass through of additional data to newly created bodies
using kwargs.