Define absolute tolerance when comparing with 0.0
By default, assert_quantity_allclose() allows only a relative
tolerance (1e-7), while the absolute tolerance is zero. For comparison
with 0.0, this is however useless, and leads to error on some CPU
types (arm64, i386).

The value 1e-12 is chosen ad-hoc to let the test pass; it may need an
adjustment.