Fixed 2 tests for spheroid location, fixed mistake in cartesian_to_ellipsoidal
In case of lat == 0 the cartesian_to_ellipsoidal the code returned
a wrong value. On inspection the code used x instead of np.sqrt(x**2 + y**2) in the formula, of which the latter should be used according to  R. Burtch, A COMPARISON OF METHODS USED IN RECTANGULAR TO GEODETIC COORDINATE TRANSFORMATIONS, page 2 and page 9.